qemu 1.2.2 release, upstream tag v1.2.2, Dec 11, 2012
with the one commit omitted just like the last one, 1.3.0 has the fix. 0b57e287

+++ qemu/hw/qdev-properties.c
@@ -913,7 +913,7 @@ static void set_blocksize(Object *obj, Visitor *v, void *opaque,
     /* We rely on power-of-2 blocksizes for bitmasks */
     if ((value & (value - 1)) != 0) {
         error_set(errp, QERR_PROPERTY_VALUE_NOT_POWER_OF_2,
-                  dev->id ? dev->id :"", name, value);
+                  dev->id ? dev->id :"", name, (int64_t)value);
         return;
     }
